refactor(cvc): extract create-only write into shared firestore-write.js
