+++ clinvar-cvc/popup.html
@@ -178,6 +178,7 @@
   <script src="popup-view.js"></script>
   <script src="history.js"></script>
   <script src="firestore-history.js"></script>
+  <script src="firestore-write.js"></script>
   <script src="popup.js"></script>
 </body>
 </html>
